restructure modules for more flexibility
I split the module into broker, protocol, and client.
For the reusability and flexibility.
