+++ broker/src/main/java/org/dsngroup/broke/broker/ServerContext.java
@@ -14,8 +14,14 @@
  * limitations under the License.
  */
 
-/**
- * This package used for dealing with channel, as a pipeline of processing.
- */
+package org.dsngroup.broke.broker;
 
-package org.dsngroup.broke.channel;
+/**
+ * The ServerContext class records the associated information of a Server.
+ */
+public class ServerContext {
+
+    // TODO: The ServerContext constructs the associated information of a Server,
+    // will be used depends on what we need in the future.
+    public ServerContext() {}
+}
