+++ broker/src/main/java/org/dsngroup/broke/broker/channel/handler/package-info.java
@@ -18,4 +18,4 @@
  * This package used for handler, for processing on channel.
  */
 
-package org.dsngroup.broke.channel.handler;
+package org.dsngroup.broke.broker.channel.handler;
