+++ client/pom.xml
@@ -4,10 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.dsngroup</groupId>
-    <artifactId>broke</artifactId>
+    <groupId>org.dsngroup.broke</groupId>
+    <artifactId>client</artifactId>
     <version>0.1</version>
-
+    <properties>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+    </properties>
     <dependencies>
         <dependency>
             <groupId>io.netty</groupId>
@@ -15,6 +18,12 @@
             <version>4.1.16.Final</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>org.dsngroup.broke</groupId>
+            <artifactId>protocol</artifactId>
+            <version>0.1</version>
+        </dependency>
     </dependencies>
+
 
 </project>
