+++ protocol/src/main/java/org/dsngroup/broke/protocol/CriticalOption.java
