+++ protocol/src/main/java/org/dsngroup/broke/protocol/Message.java
